+++ lib/tableau_api.rb
@@ -13,6 +13,7 @@ require 'tableau_api/resources/users'
 require 'tableau_api/resources/groups'
 require 'tableau_api/resources/workbooks'
 require 'tableau_api/resources/datasources'
+require 'tableau_api/resources/jobs'
 require 'tableau_api/version'
 
 module TableauApi