+++ lib/tableau_api/resources/jobs.rb
@@ -1,0 +1,10 @@
+module TableauApi
+  module Resources
+    class Jobs < Base
+      def list(kwargs = {})
+        url = "sites/#{@client.auth.site_id}/jobs"
+        @client.connection.api_get_collection(url, 'backgroundJobs.backgroundJob', **kwargs)
+      end
+    end
+  end
+end