+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] -
+
+### Added
+
+- Added Jobs resource
+
+
+### Changed
+
+- Updated to API version 3.1
+  - This is a breaking change for site roles: 
+    https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_new_site_roles.htm
+- Include full error detail in TableauError message
+
+
 ## [2.0.0] - 2019-08-29
 
 - Updated to API version 2.8, compatible with Tableau Server >= 10.5